WIP: Updated bigint to return long long

+++ src/algebra/fields/bigint.tcc
@@ -105,7 +105,7 @@ bool bigint<n>::is_zero() const
 template<mp_size_t n>
 size_t bigint<n>::num_bits() const
 {
-/*
+
     for (long i = max_bits(); i >= 0; --i)
     {
         if (this->test_bit(i))
@@ -115,7 +115,7 @@ size_t bigint<n>::num_bits() const
     }
 
     return 0;
-*/
+/*
     for (long i = n-1; i >= 0; --i)
     {
         mp_limb_t x = this->data[i];
@@ -129,10 +129,11 @@ size_t bigint<n>::num_bits() const
         }
     }
     return 0;
+*/
 }
 
 template<mp_size_t n>
-unsigned long bigint<n>::as_ulong() const
+unsigned long long bigint<n>::as_ulong() const
 {
     return this->data[0];
 }
